+++ src/checkout/types.ts
@@ -1,7 +1,7 @@
 import type { ContractFunctionParameters, TransactionReceipt } from 'viem';
 import type { Address } from 'viem';
 import type { Config } from 'wagmi';
-import type { PayTransaction } from '../api/types';
+import type { PayTransaction } from '../packages/core/api/types';
 import type { TransactionError } from '../transaction';
 
 /**
