+++ src/nft/hooks/useMintDetails.test.ts
@@ -1,10 +1,10 @@
 import { renderHook, waitFor } from '@testing-library/react';
 import { type Mock, beforeEach, describe, expect, it, vi } from 'vitest';
-import { getMintDetails } from '../../api/getMintDetails';
+import { getMintDetails } from '../../packages/core/api/getMintDetails';
 import { getNewReactQueryTestProvider } from '../../identity/hooks/getNewReactQueryTestProvider';
 import { useMintDetails } from './useMintDetails';
 
-vi.mock('../../api/getMintDetails');
+vi.mock('../../packages/core/api/getMintDetails');
 
 describe('useMintDetails', () => {
   beforeEach(() => {
