+++ src/nft/hooks/useTokenDetails.ts
@@ -1,6 +1,9 @@
 import { type UseQueryResult, useQuery } from '@tanstack/react-query';
-import { getTokenDetails } from '../../api/getTokenDetails';
-import type { GetTokenDetailsParams, TokenDetails } from '../../api/types';
+import { getTokenDetails } from '../../packages/core/api/getTokenDetails';
+import type {
+  GetTokenDetailsParams,
+  TokenDetails,
+} from '../../packages/core/api/types';
 import { isNFTError } from '../utils/isNFTError';
 
 export function useTokenDetails({
