+++ src/packages/core/api/buildPayTransaction.ts
@@ -1,8 +1,8 @@
 import {
   CDP_CREATE_PRODUCT_CHARGE,
   CDP_HYDRATE_CHARGE,
-} from '../network/definitions/pay';
-import { type JSONRPCResult, sendRequest } from '../network/request';
+} from '../../../network/definitions/pay';
+import { type JSONRPCResult, sendRequest } from '../../../network/request';
 import type {
   BuildPayTransactionParams,
   BuildPayTransactionResponse,
