+++ src/packages/core/api/types.ts
@@ -1,7 +1,12 @@
 import type { Address } from 'viem';
-import type { ContractType, NFTPrice } from '../nft/types';
-import type { Fee, QuoteWarning, SwapQuote, Transaction } from '../swap/types';
-import type { Token } from '../token/types';
+import type { ContractType, NFTPrice } from '../../../nft/types';
+import type {
+  Fee,
+  QuoteWarning,
+  SwapQuote,
+  Transaction,
+} from '../../../swap/types';
+import type { Token } from '../../../token/types';
 
 export type AddressOrETH = Address | 'ETH';
