+++ src/packages/core/api/utils/getPayErrorMessage.test.ts
@@ -5,7 +5,7 @@ import {
   CHECKOUT_TOO_MANY_REQUESTS_ERROR_MESSAGE,
   GENERAL_CHECKOUT_ERROR_MESSAGE,
   UNCAUGHT_CHECKOUT_ERROR_MESSAGE,
-} from '../../checkout/constants';
+} from '../../../../checkout/constants';
 import { getPayErrorMessage } from './getPayErrorMessage';
 
 describe('getPayErrorMessage', () => {
